Module type Mirage_random.S

Operations to generate random numbers

type g

The state of the generator.

val generate : ?g:g -> int -> string

generate ~g len generates a random buffer of length len using g.

val generate_into : ?g:g -> bytes -> ?off:int -> int -> unit

generate_into ~g buf ~off len puts len bytes random data into buf starting at off (defaults to 0).